+ support for the "unique(all:true)" case
